Feat: 알림창 추억회상 친구추가
알림창 추억회상...

+++ server/photomory/src/main/java/com/example/photomory/repository/MyAlbumRepository.java
@@ -3,5 +3,5 @@ package com.example.photomory.repository;
 import com.example.photomory.entity.MyAlbum;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MyAlbumRepository extends JpaRepository<MyAlbum, Integer> {
+public interface MyAlbumRepository extends JpaRepository<MyAlbum, Long> {
 }
